Add Generic Dependency aware processor implementation
This processor implementation handles the notion of dependencies between processors, such that a processor A which depends on processor B can only be enabled is B is enabled as well. In case A is enabled but B is not, an info logging message is displayed stating that A cannot be enabled due to B's disablement.